refractor: camera control successfully refractored and tested
