+++ obsbot_cli/commands/camera_controls/__init__.py
@@ -1,0 +1,4 @@
+# commands/camera_controls/__init__.py
+from .controller import CameraControlCommands
+
+__all__ = ['CameraControlCommands']
