+++ obsbot_cli/commands/camera_controls/utils/__init__.py
@@ -1,0 +1,6 @@
+# commands/camera_controls/utils/__init__.py
+# Make conversion and movement utilities available at package level
+from .conversion import AngleConverter
+from .movement import MovementHandler
+
+__all__ = ['AngleConverter', 'MovementHandler']
